+++ src/error.rs
@@ -3,4 +3,8 @@
 pub enum RextTuiError {
     #[error("Failed to read crossterm event: {0}")]
     ReadEvent(#[from] std::io::Error),
+    #[error("Failed to load config: {0}")]
+    ConfigError(#[from] toml::de::Error),
+    #[error("Failed to read config file: {0}")]
+    ReadConfigFile(std::io::Error),
 }